Fix arm, armhf and armv5te and missing `__atomic`
Because of some quirks during linking sometimes the `__atomic` isn't
found or linked or something by the rustc compiler.

Because of this we need to add this via `RUSTFLAGS`.
Added this to the `test.sh` script, the GHA Workflow and the README.

This is triggered because of adding the `mimalloc` crate to the `multicrate` for testing.
But, since this is a real world example (used by Vaultwarden for example),
and using MiMalloc as an allocator is very useful for musl based
application, I have added this as a default to catch issues (like this).