Update base image to be able to run bindgen
Since Diesel v2.2.x and mysqlclient-sys v0.3.x it is possible (and
sometimes needed) to generate bindgen information. For this to be able
to work correctly I needed to add the `linclang-dev` package to be
installed.

Also, I noticed a bug in the final `ENV` variables for
`TARGET_C_INCLUDE_PATH` and i needed to add some extra arguments for
`BINDGEN_EXTRA_CLANG_ARGS`.

This together with a yet to be pushed/merged change for mysqlclient-sys
to actually allow the bindgen to work on non-x86_64 architecture should
fix building the newer version of Diesel/mysqlclient-sys as for any
other library which needs to do this in the future :).

Also, bindgen needed `rustfmt`, I also added this to be part of the
install to prevent any issues.